neonvm-runner: Move http server to its own file (#1185)
